+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Corrected code-smell and code style issues in `\ncc\Classes > ComposerExtension > ComposerSourceBuiltin`
  - Corrected code-smell and code style issues in `\ncc\Classes > GithubExtension > GithubService`
  - Corrected code-smell and code style issues in `\ncc\Classes > GitlabExtension > GitlabService`
+ - Corrected code-smell and code style issues in `\ncc\Classes > NccExtension > ConstantCompiler`